Bug Fix, Updated Control Versatility, Finished Primative Points System
Point system that was said to have been implemented previously was not fully finished. The primative point system works the same way as before, but the earlier commit did not have all the work. The point system is now fully impelmented in this version.

Updated the amount of string maps in the control scheme to allow for more keys to be bound in the game

Fixed a bug where selecting the control buttons in the options menu did not correspond with the actual buttons rendered.