Full Circle
This commit comes with tons of new additions to the game. This is mainly a
staging commit to save the progress made so far in case something happens
to our own copy. In this case we have lots of commented out code which is
our experimentation with audio processing and different features of the
game. Each menu system is being used to test out new features of the game
and is not working properly. Finally, a small particle effect has been
added to the circles when hit and the points system has broken down.
Circle collision detction had to be changed in order to inplement the
particle system and a new class was added to deal with the particles.
All changes may not be covered in this commit message due to the nature of
this commit as a staging commit.