Points System
Points system almost fully implemented. At the moment there is a basic point system working which will give you points based on how close you hit the circles to the middle of the hitbox. Along with this, we are testing a new menu system in the tutorial and arcade states, which are still under development. We have a story engine separate at the moment, but will be moving it into the game once we know the theme for OGPC.