Safe Commit, Updated controls
We have managed to find out the theme for this years OGPC competition.
Turns out our game is going to be about mythology. We are working out a
story mode now, but I is not incorporated into this version of the game
at the moment. This commit is mainly to make some small changes and to
have a safe place to fall back to. I am going to start redoing the game
engine and state. I want to clean them up a bit so I need to mess around
with a lot of the code.
Along with safe committing I have also moved the game back to the 5
radar circles that it originally had. We have decided to change the
control scheme. We will be creating a dance pad for the OGPC
competition. Three buttons (to work within the control restrictions that
they originally had placed and then removed) are used to control the
three circles. Button combinations are then used to hit the middle
circles. Currently the buttons are set to h, j, and k keys. New control
scheme is no longer mapable because we will be using a dance pad as our
control scheme.