Updated Engine and Arcade State
Reprogramed the rhythm engine in order to make the rhythms sync up with
music. Arcade state is now almost fully operational with only minor
touch ups needed (sampling music).  Basic control scheme for the moment
is set to the buttons H, J, and K. We have some arduino projects in the
works currently, which will help us make the move to a dance pad. Basic
game is running and though only the arcade state is fully operational.
Next task is to redo the menu system. The menu system will most likely
get almost a complete overhaul in the next commit.