Updated Menu System, Added Sound, and Added Story
Big new update. This update comes with a completely redone menu system
which works similar to how the recently updated arcade state works.
Three main buttons are still h, j, and k, use these to control the menu
system and move in between states. I have also added into the game sound
effects and some music for the menu screen. These are just testing at
the moment, but some of them (especially the mainmenu music may be here
to stay if we can't find anything better. Along with this the story
engine, which we were working on separately has finally been ported into
the game. As of yet, it is not working fully, but we hope to have it up
and running in the next couple of weeks. Next up on our tasks are to
come up with a game story, design our controller, and work on the
engine. Next update will most likely involve a sample story and better
actual gameplay.