Logo Update and A Safe Commit
This update comes with very little features, in fact, you will see
almost no difference between this update an the last. The main point of
this update is to serve as a backup. Along with the backup this update
also comes with two new classes. Though not functioning in the game as
of yet, we have created a new logo for full circle the game and are
changing the current logo to the Team Henry logo. We also (while in the
process of creating our new logo) inadvertently created a new loading
screen which we will implement in the next update. For now, the game
remains much the same. Next update will come with two logos, one for
Team Henry and the new updated full circle one.