Full Logo Update and Color Change
New update to the aesthetics of the gameplay. We recently added a team
henry logo along with the new Full Circle logo. When you start the game
you should get both logos (press escape to skip them). We also changed
our games color scheme. The three primary colors are no long the simple
rgb values that they were before, the are now more muted. Finally, we
also upgraded the menu music to be more DDR-like and are in the process
of coming up with and implementing a story. The next commit will most
likely involve new gameplay mechanics and/or the story being fully
implemented into the game.