Story Update and Another Safe Commit
We added a sample storyline to the game. Simple text files with very
little substance. Story state now works properly moving from story to
rhythm and back to story again. I also added a story state to the game
as well. The commit is mainly going to be used for safety as the rhythm
engine generation and display needs to be messed with. We are moving
away from angle timing and moving towards song timings.